Restructure vm
