+++ .gitignore
@@ -18,3 +18,8 @@ MANIFEST-*
 OPTIONS-*
 CURRENT
 LOCK
+
+# Generated by Cargo
+# will have compiled files and executables
+/target/
+Cargo.lock
