+++ Cargo.toml
@@ -1,0 +1,25 @@
+[workspace]
+members = ["vm/rust", "starknet/compiler/rust", "core/rust"]
+resolver = "2"
+
+[workspace.package]
+version = "0.0.0"
+edition = "2021"
+rust-version = "1.81"
+
+[workspace.dependencies]
+serde = "1.0.208"
+serde_json = { version = "1.0.125", features = ["raw_value"] }
+blockifier = "=0.14.0-rc.0"
+cairo-lang-starknet-classes = "=2.10.0-rc.1"
+cairo-lang-runner = { version = "2.10.0-rc.0" }
+starknet_api = "=0.14.0-rc.0"
+cairo-vm = "=1.0.1"
+starknet-types-core = { version = "0.1.6", features = ["hash", "prime-bigint"] }
+indexmap = "2.1.0"
+cached = "0.54.0"
+once_cell = "1.19.0"
+lazy_static = "1.4.0"
+semver = "1.0.22"
+anyhow = "1.0.81"
+starknet-core = "0.8.0"
