+++ core/rust/Cargo.toml
@@ -1,14 +1,15 @@
 [package]
 name = "juno-starknet-core-rs"
-version = "0.0.0"
-edition = "2021"
+version = { workspace = true }
+edition = { workspace = true }
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-serde = "1.0.171"
-serde_json = { version = "1.0.96", features = ["raw_value"] }
-starknet-core = "0.8.0"
+serde = { workspace = true }
+serde_json = { workspace = true, features = ["raw_value"] }
+starknet-core = { workspace = true }
 
 [lib]
 crate-type = ["staticlib"]
+
