+++ rustfmt.toml
@@ -1,0 +1,11 @@
+format_code_in_doc_comments = true
+normalize_doc_attributes = true
+wrap_comments = true
+
+format_strings = true
+
+use_field_init_shorthand = true
+
+imports_layout = "HorizontalVertical"
+group_imports = "StdExternalCrate"
+imports_granularity = "Module"
