+++ vm/rust/Cargo.toml
@@ -6,20 +6,20 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-serde = "1.0.208"
-serde_json = { version = "1.0.125", features = ["raw_value"] }
-blockifier = "=0.14.0-rc.0"
-cairo-lang-starknet-classes = "=2.10.0-rc.1"
-cairo-lang-runner = { version = "2.10.0-rc.0" }
-starknet_api = "=0.14.0-rc.0"
-cairo-vm = "=1.0.1"
-starknet-types-core = { version = "0.1.6", features = ["hash", "prime-bigint"] }
-indexmap = "2.1.0"
-cached = "0.54.0"
-once_cell = "1.19.0"
-lazy_static = "1.4.0"
-semver = "1.0.22"
-anyhow = "1.0.81"
+serde = { workspace = true }
+serde_json = { workspace = true, features = ["raw_value"] }
+blockifier = { workspace = true }
+cairo-lang-starknet-classes = { workspace = true }
+cairo-lang-runner = { workspace = true }
+starknet_api = { workspace = true }
+cairo-vm = { workspace = true }
+starknet-types-core = { workspace = true, features = ["hash", "prime-bigint"] }
+indexmap = { workspace = true }
+cached = { workspace = true }
+once_cell = { workspace = true }
+lazy_static = { workspace = true }
+semver = { workspace = true }
+anyhow = { workspace = true }
 
 [lib]
 crate-type = ["staticlib"]
