+++ vm/rust/src/errors.rs
@@ -1,0 +1,9 @@
+use crate::ffi::JunoReportError;
+use std::ffi::CString;
+
+pub fn report_error(reader_handle: usize, msg: &str, txn_index: i64) {
+    let err_msg = CString::new(msg).unwrap();
+    unsafe {
+        JunoReportError(reader_handle, txn_index, err_msg.as_ptr());
+    };
+}
