+++ vm/vm.go
@@ -1,8 +1,8 @@
 package vm
 
 /*
-#cgo vm_debug  LDFLAGS: -L./rust/target/debug   -ljuno_starknet_rs -lbz2
-#cgo !vm_debug LDFLAGS: -L./rust/target/release -ljuno_starknet_rs -lbz2
+#cgo vm_debug  LDFLAGS: -L../target/debug   -ljuno_starknet_rs -lbz2
+#cgo !vm_debug LDFLAGS: -L../target/release -ljuno_starknet_rs -lbz2
 
 #include "vm_ffi.h"
 */
